Changed page context to be the tutor group context.
Previously, the page context for view_main would operate in the context
of the user being viewed.  This allowed the tutor to be identified,
but not the LAA. This changes it to use the context of the user's
tutor group, with the exception of where a user is viewing their
own PLP, allowing all roles to be identified.